move Cnossos attenuation stuff from pathfinder to propagation package
