+++ noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/PathFinderVisitor.java
@@ -8,43 +8,37 @@
  */
 package org.noise_planet.noisemodelling.pathfinder;
 
-
-import org.h2gis.api.EmptyProgressVisitor;
-import org.h2gis.api.ProgressVisitor;
-import org.noise_planet.noisemodelling.pathfinder.cnossos.CnossosPath;
 import org.noise_planet.noisemodelling.pathfinder.path.Scene;
+import org.noise_planet.noisemodelling.pathfinder.profilebuilder.CutProfile;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
+import java.util.Collection;
+import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Way to store data computed by threads.
  * Multiple threads use one instance.
  * This class must be thread safe
- * Store only propagation rays
+ * Store only vertical cut planes
  * @author Nicolas Fortin
  * @author Pierre Aumond
  */
 public class PathFinderVisitor implements IComputePathsOut {
-    public List<CnossosPath> pathParameters = Collections.synchronizedList(new ArrayList<>());
+    /** This list is thread safe so can be used in a multi-thread environment */
+    public ConcurrentLinkedDeque<CutProfile> cutProfiles = new ConcurrentLinkedDeque<>();
     public Scene inputData;
 
-    public PathFinderVisitor(boolean keepRays, Scene inputData) {
-        this.keepRays = keepRays;
+    public boolean keepCutPlanes = true;
+    public AtomicLong pathCount = new AtomicLong();
+
+    public PathFinderVisitor(boolean keepCutPlanes, Scene inputData) {
+        this.keepCutPlanes = keepCutPlanes;
         this.inputData = inputData;
     }
 
-    public PathFinderVisitor(boolean keepRays) {
-        this.keepRays = keepRays;
+    public PathFinderVisitor(boolean keepCutPlanes) {
+        this.keepCutPlanes = keepCutPlanes;
     }
-
-    public boolean keepRays = true;
-    public AtomicLong rayCount = new AtomicLong();
 
     /**
      * No more propagation paths will be pushed for this receiver identifier
@@ -60,19 +54,13 @@ public class PathFinderVisitor implements IComputePathsOut {
     }
 
 
-    /**
-     * Get propagation path result
-     * @param sourceId Source identifier
-     * @param sourceLi Source power per meter coefficient
-     * @param path Propagation path result
-     */
     @Override
-    public double[] addPropagationPaths(long sourceId, double sourceLi, long receiverId, List<CnossosPath> path) {
-        rayCount.addAndGet(path.size());
-        if (keepRays) {
-            pathParameters.addAll(path);
+    public PathSearchStrategy onNewCutPlane(CutProfile cutProfile) {
+        pathCount.addAndGet(1);
+        if (keepCutPlanes) {
+            cutProfiles.add(cutProfile);
         }
-        return new double[0];
+        return PathSearchStrategy.CONTINUE;
     }
 
     /**
@@ -81,74 +69,10 @@ public class PathFinderVisitor implements IComputePathsOut {
      */
     @Override
     public IComputePathsOut subProcess() {
-        return new ThreadPathsOut(this);
+        return this;
     }
 
-    public List<CnossosPath> getPropagationPaths() {
-        return pathParameters;
+    public Collection<CutProfile> getCutProfiles() {
+        return cutProfiles;
     }
-
-    public static class ThreadPathsOut implements IComputePathsOut {
-        protected PathFinderVisitor multiThreadParent;
-        public List<CnossosPath> pathParameters = new ArrayList<>();
-
-        public ThreadPathsOut(PathFinderVisitor multiThreadParent) {
-            this.multiThreadParent = multiThreadParent;
-        }
-
-        /**
-         * Get propagation path result
-         * @param sourceId Source identifier
-         * @param sourceLi Source power per meter coefficient
-         * @param path path result
-         */
-        @Override
-        public double[] addPropagationPaths(long sourceId, double sourceLi, long receiverId, List<CnossosPath> path) {
-            multiThreadParent.rayCount.addAndGet(path.size());
-            if (multiThreadParent.keepRays) {
-                if (multiThreadParent.inputData != null && sourceId < multiThreadParent.inputData.sourcesPk.size() &&
-                        receiverId < multiThreadParent.inputData.receiversPk.size()) {
-                    for (CnossosPath pathParameter : path) {
-                        pathParameter.setIdReceiver(multiThreadParent.inputData.receiversPk.get((int) receiverId).intValue());
-                        pathParameter.setIdSource(multiThreadParent.inputData.sourcesPk.get((int) sourceId).intValue());
-                        pathParameters.add(pathParameter);
-                    }
-                } else {
-                    pathParameters.addAll(path);
-                }
-            }
-            return new double[0];
-        }
-
-        /**
-         * No more propagation paths will be pushed for this receiver identifier
-         * @param receiverId
-         */
-    @Override
-    public void finalizeReceiver(final long receiverId) {
-        if (multiThreadParent.keepRays && !pathParameters.isEmpty()) {
-            multiThreadParent.pathParameters.addAll(pathParameters);
-            pathParameters.clear();
-        }
-        long receiverPK = receiverId;
-        if (multiThreadParent.inputData != null) {
-            if (receiverId < multiThreadParent.inputData.receiversPk.size()) {
-                receiverPK = multiThreadParent.inputData.receiversPk.get((int) receiverId);
-            }
-        }
-        multiThreadParent.finalizeReceiver(receiverId);
-
-    }
-
-        /**
-         *
-         * @return an instance of the interface IComputePathsOut
-         */
-
-    @Override
-    public IComputePathsOut subProcess() {
-        return multiThreadParent.subProcess();
-    }
-}
-
 }
